--- autopush/e2e-test/tests/pop-notification-GR32.sikuli/pop-notification-GR32.html
@@ -1,97 +1,0 @@
-
-<html>
-   <head>
-      <style type="text/css">
-         .sikuli-code {
-            font-size: 20px;
-            font-family: "Osaka-mono", Monospace;
-            line-height: 1.5em;
-            display:table-cell;
-            white-space: pre-wrap;       /* css-3 */
-            white-space: -moz-pre-wrap !important;  /* Mozilla, since 1999 */
-            white-space: -pre-wrap;      /* Opera 4-6 */
-            white-space: -o-pre-wrap;    /* Opera 7 */
-            word-wrap: break-word;       /* Internet Explorer 5.5+ */
-            width: 99%;   /* remove horizontal scroll-bar when viewing in IE7 */
-         }
-         .sikuli-code img {
-            vertical-align: middle;
-            margin: 2px;
-            border: 1px solid #ccc;
-            padding: 2px;
-            -moz-border-radius: 5px;
-            -webkit-border-radius: 5px;
-            -moz-box-shadow: 1px 1px 1px gray;
-            -webkit-box-shadow: 1px 1px 2px gray;
-         }
-         .kw {
-            color: blue;
-         }
-         .skw {
-            color: rgb(63, 127, 127);
-         }
-
-         .str {
-            color: rgb(128, 0, 0);
-         }
-
-         .dig {
-            color: rgb(128, 64, 0);
-         }
-
-         .cmt {
-            color: rgb(200, 0, 200);
-         }
-
-         h2 {
-            display: inline;
-            font-weight: normal;
-         }
-
-         .info {
-            border-bottom: 1px solid #ddd;
-            padding-bottom: 5px;
-            margin-bottom: 20px;
-            display: none;
-         }
-
-         a {
-            color: #9D2900;
-         }
-
-         body {
-            font-family: "Trebuchet MS", Arial, Sans-Serif;
-         }
-
-      </style>
-   </head>
-<body>
-<div class="info">
-<h2>pop-notification-GR32.sikuli</h2> <a href="pop-notification-GR32.zip">(Download this script)</a>
-</div>
-<pre class="sikuli-code">
-NIGHTLY64 = <span class="str">"C:/Program Files/Nightly/firefox.exe"</span>
-GENREL32 = <span class="str">"C:/Program Files (x86)/Mozilla Firefox/firefox.exe"</span>
-
-
-App.open(GENREL32)
-
-<span class="skw">click</span>(<img src="1446153249015.png" />)
-
-<span class="skw">sleep</span>(<span class="dig">2</span>)
-
-<span class="kw">if</span> exists(<img src="1446153637218.png" />):
-    <span class="skw">click</span>(<img src="1446153637218.png" />)
-
-<span class="kw">print</span> <span class="str">"here?"</span>
-
-<span class="skw">wait</span>(<img src="1446153338359.png" />)
-
-waitVanish(<img src="1446153338359.png" />)
-
-<span class="kw">print</span> <span class="str">"yay"</span>
-
-<span class="skw">click</span>(<img src="1446154091312.png" />)
-</pre>
-</body>
-</html>
