--- autopush/e2e-test/tests/pop-notification-NN64.sikuli/pop-notification-NN64.html
@@ -1,158 +1,0 @@
-
-<html>
-   <head>
-      <style type="text/css">
-         .sikuli-code {
-            font-size: 20px;
-            font-family: "Osaka-mono", Monospace;
-            line-height: 1.5em;
-            display:table-cell;
-            white-space: pre-wrap;       /* css-3 */
-            white-space: -moz-pre-wrap !important;  /* Mozilla, since 1999 */
-            white-space: -pre-wrap;      /* Opera 4-6 */
-            white-space: -o-pre-wrap;    /* Opera 7 */
-            word-wrap: break-word;       /* Internet Explorer 5.5+ */
-            width: 99%;   /* remove horizontal scroll-bar when viewing in IE7 */
-         }
-         .sikuli-code img {
-            vertical-align: middle;
-            margin: 2px;
-            border: 1px solid #ccc;
-            padding: 2px;
-            -moz-border-radius: 5px;
-            -webkit-border-radius: 5px;
-            -moz-box-shadow: 1px 1px 1px gray;
-            -webkit-box-shadow: 1px 1px 2px gray;
-         }
-         .kw {
-            color: blue;
-         }
-         .skw {
-            color: rgb(63, 127, 127);
-         }
-
-         .str {
-            color: rgb(128, 0, 0);
-         }
-
-         .dig {
-            color: rgb(128, 64, 0);
-         }
-
-         .cmt {
-            color: rgb(200, 0, 200);
-         }
-
-         h2 {
-            display: inline;
-            font-weight: normal;
-         }
-
-         .info {
-            border-bottom: 1px solid #ddd;
-            padding-bottom: 5px;
-            margin-bottom: 20px;
-            display: none;
-         }
-
-         a {
-            color: #9D2900;
-         }
-
-         body {
-            font-family: "Trebuchet MS", Arial, Sans-Serif;
-         }
-
-      </style>
-   </head>
-<body>
-<div class="info">
-<h2>pop-notification-NN64.sikuli</h2> <a href="pop-notification-NN64.zip">(Download this script)</a>
-</div>
-<pre class="sikuli-code">
-APP_NAME = <span class="str">'Nightly'</span>
-<span class="cmt">#APP_NAME = 'Release'
-</span>NIGHTLY64 = <span class="str">"C:/Program Files/%s/firefox.exe -height 768 -width 1024"</span> % APP_NAME
-PATH_FIREFOX = NIGHTLY64
-
-<span class="kw">if</span> APP_NAME == <span class="str">'Nightly'</span>:
-    PATH_IMGS = <span class="str">'../images_nightly64/'</span>
-<span class="kw">else</span>:
-    PATH_IMGS = <span class="str">'../images_release32/'</span>
-    <span class="kw">print</span>(<span class="str">'TBD -&gt; need images for GR32!!'</span>)
-    exit(<span class="dig">1</span>)
-
-
-URL_TEST_PAGE = <span class="str">'https://pdehaan.github.io/push-notification-test'</span>
-
-IMG_REFRESH = PATH_IMGS + <img src="btn_refresh.png" /> <span class="cmt"># </span>
-URL_ENTRY_POINT_1 = Pattern(PATH_IMGS + <img src="url_entry_moz.png" />).targetOffset(<span class="dig">110</span>,-<span class="dig">4</span>)
-URL_ENTRY_POINT_2 = PATH_IMGS + <img src="url_entry_search.png" />
-URL_ENTRY_POINT_3 = Pattern(PATH_IMGS + <img src="url_entry_nightly.png" />).targetOffset(<span class="dig">59</span>,<span class="dig">0</span>)
-URL_ENTRY_POINT_4 = Pattern(PATH_IMGS + <img src="url_entry_green_padlock.png" />).targetOffset(<span class="dig">59</span>,<span class="dig">1</span>)
-IMG_WIN_MAXIMIZE = PATH_IMGS + <img src="win_maximize.png" />
-IMG_ALWAYS_RECEIVE_NOTIFICATIONS = PATH_IMGS + <img src="btn_always_receive.png" />
-IMG_BTN_POP_NOTIFICATION = PATH_IMGS + <img src="btn_pop_notification.png" />
-IMG_WIN_POP_NOTIFICATION = PATH_IMGS + <img src="win_pop_notification.png" />
-LINE = <span class="str">'--------------------------'</span>
-
-
-firefox = App(APP_NAME)
-
-<span class="kw">def</span> header(label):
-    <span class="kw">return</span> <span class="str">'%s\n%s\n%s'</span> % (LINE, label, LINE)
-
-<span class="kw">def</span> setup():
-
-    <span class="kw">print</span> header(<span class="str">'SETUP'</span>)
-    firefox.open(PATH_FIREFOX)
-    <span class="kw">if</span> exists(IMG_REFRESH):
-        <span class="skw">click</span>(IMG_REFRESH)
-        <span class="skw">click</span>(IMG_REFRESH)
-    firefox.focus()
-
-<span class="kw">def</span> teardown():
-
-    <span class="cmt">#print('%s\n%s\n%s' % (LINE, 'TEARDOWN', LINE))
-</span>    <span class="kw">print</span> header(<span class="str">'TEARDOWN'</span>)
-    <span class="kw">print</span>(<span class="str">'closing browser now'</span>)
-    firefox.close()
-    <span class="kw">print</span>(<span class="str">'Firefox closed'</span>)
-
-setup()
-
-<span class="kw">print</span> header(<span class="str">'TEST'</span>)
-
-<span class="kw">if</span> exists(URL_ENTRY_POINT_1):
-    <span class="kw">print</span>(<span class="str">'URL ENTRY POINT: &lt;Mozilla Foundation&gt;'</span>)
-    <span class="skw">click</span>(URL_ENTRY_POINT_1)
-<span class="kw">elif</span> exists(URL_ENTRY_POINT_2):
-    <span class="kw">print</span>(<span class="str">'URL ENTRY POINT: &lt;Search or enter address&gt;'</span>)
-    <span class="skw">click</span>(URL_ENTRY_POINT_2)
-<span class="kw">elif</span> exists(URL_ENTRY_POINT_3):
-    <span class="kw">print</span>(<span class="str">'URL ENTRY POINT: &lt;Nightly...&gt;'</span>)
-    <span class="skw">click</span>(URL_ENTRY_POINT_3)
-<span class="kw">else</span>:
-    <span class="kw">print</span>(<span class="str">'URL ENTRY POINT: &lt;green padlock&gt;'</span>)
-    <span class="skw">click</span>(URL_ENTRY_POINT_4)
-<span class="skw">type</span>(URL_TEST_PAGE + Key.ENTER)
-
-<span class="skw">click</span>(IMG_BTN_POP_NOTIFICATION)
-
-<span class="skw">sleep</span>(<span class="dig">2</span>)
-
-<span class="kw">print</span>(<span class="str">'PERMISSIONS: &lt;Always Receive Notifications?&gt;'</span>)
-<span class="kw">if</span> exists(IMG_ALWAYS_RECEIVE_NOTIFICATIONS):
-    <span class="skw">click</span>(IMG_ALWAYS_RECEIVE_NOTIFICATIONS)
-
-<span class="kw">print</span> <span class="str">"POP NOTIFICATION: ???"</span>
-<span class="skw">wait</span>(IMG_WIN_POP_NOTIFICATION)
-<span class="kw">print</span>(<span class="str">'POP NOTIFICATION: COMPLETE!'</span>)
-<span class="kw">print</span>(<span class="str">'POP NOTIFICATION: WAITING FOR VANISH.....'</span>)
-waitVanish(IMG_WIN_POP_NOTIFICATION)
-<span class="kw">print</span>(<span class="str">'TEST COMPLETE!'</span>)
-
-teardown()
-</pre>
-</body>
-</html>
